Fix to #3633 - Query :: QuerySourceTracingExpressionVisitor doesn't prune EF.Property calls when looking for query sources
QuerySourceTracingExpressionVisitor is looking for query sources that are being materialized, so that we know which includes need to be applied to the query. The logic filters out references that are not being projected (e.g. property access), however we were not pruning property access done via EF.Property method.

Fix is to add the missing overload.

CR: Andrew