Fix to #3616 - Projection with multiple aggregates only running last one
Problem was that when we were adding a subquery to the projection we were clearing everything that was in the projection list before. This is not a problem for scalars, because they are all re-added when we bind the query, but complex expressions are not added and are ignored. Fix is to only remove column projection elements instead.